add mise for simple dev environments and task management
most of my difficulty in contributing to radarr stemmed from simply
getting my development environment setup.

we should prioritize tooling to make the developer setup as painless as
possible in order to make it as easy as possible for people from all
backgrounds to contribute.

i'm using `mise` here because it's not tied to a particular language
ecosystem; it's lightweight and doesn't require containerization or
root; it supports separate "dev" and "prod" environments; and it adds a
convenient task system on top of `asdf`.