keep routes identical
we want the metadata server to be configurable, but we will also have a
twilight period where some users still need to use GR while others want
a different server.

let's try to stay as close as possible to the existing GR routes. this
will make it easy to drop in a custom server without needing to make
changes in the client.

(this is only a temporary consideration -- we won't be able to maintain
compatibility once we start refactoring the client to take advantage of
more server-side functionality.)

note this has no impact on existing DBs. users won't be able to use new
metadata alongside existing GR IDs (i don't think...)