configure the index before ingesting
index parameters should be defined upfront so we don't waste cycles
building an index for a field that doesn't actually matter.

this isn't a big deal right now, but we will want to start returning a
lot more information from this endpoint to reduce the number of calls
the client needs to make. that will involve storing a lot more data in
each document, with most fields being ignored for search and indexing.