work with a subset of the corpus in development
this still pulls down the full dump but then trims it to a much smaller
corpus to speed up ingestion in dev. this will be helpful when making
index changes that require starting from scratch.

this will require a new "-prod" docker volume, so if you want to run the
full corpus you'll need to do a full ingest.

(fyi i don't see ratings getting returned.)

this also commits an encrypted key for prod, but it should still be
usable even if not decrypted.